Use legalizeFilename to preserve duplicats
 * move legalizeFilename out of desktop pkg

fix #2656
